Add integration test to verify health_check endpoint
Note that in the integration test, we use `tokio::spawn` to start our app in a
background task. What happens to that task when we shut down the app?
If we run `cargo test` multiple times, it works each time. Therefore, it looks
like the task is being shut down and the port used by the app is being freed.
Indeed, the tokio documentation tells us that when a tokio runtime is shut down,
all spawned tasks are dropped.